Adds 'static' assessment agent that reports No Evidence on blank code.
This refactors the grading path a little bit by adding a function for
static assessment and one for normal AI assessment. In the future, there
might be certain things we want to do to analyze the code either in the
place of AI or to supplement the AI.

Currently, this is all or nothing. Either it uses the static assessment
(which in this case only reports a No Evidence assessment if the code is
absolutely empty) or it uses the AI assessment.