Adds honeybadger support.
The Dockerfile is required to be in a non-slim image because honeybadger
introduces a dependency that is natively built. We will likely run into
these things, anyway.

Adds the `honeybadger` Python library to `requirements.txt`

New configuration option via env to provide the honeybadger project key
within the `HONEYBADGER_API_KEY` variable.

Sets it up to log all exceptions that are unhandled within the app.

Sets it up to log all CRITICAL logs that happen in the app.

Sets it up to log all ERROR logs (that aren't exceptions) in the app.